+++ changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.0.0]
+
+* namespace changes
+* add DBPhone database field
+* fix form field for usage in CMS
+
 ## [3.0.4]
 
 * Fix utils.js and allow setting initial country from back-end